fix(ci): broken build after picuture / lazy image move
