+++ packages/components/base/source/content-box/ContentBoxComponent.tsx
@@ -6,7 +6,7 @@ import {
   defaultRenderFn as richTextDefaultRenderFn,
 } from '../rich-text';
 import { Button } from '../button';
-import { Picture } from '../image/picture';
+import { Picture } from '../picture';
 import type { ContentBoxProps as ContentBoxSchemaProps } from './typing';
 
 export type ContentBoxProps = ContentBoxSchemaProps & {
