+++ tsconfig.json
@@ -15,12 +15,6 @@
       "@kickstartds/form/lib/*/typing": [
         "./packages/components/form/source/*/typing"
       ],
-      "@kickstartds/base/lib/picture/typing": [
-        "./packages/components/base/source/image/picture/typing"
-      ],
-      "@kickstartds/base/lib/lightbox-image/typing": [
-        "./packages/components/base/source/image/lightbox-image/typing"
-      ],
       "@kickstartds/blog/lib/news-item/typing": [
         "./packages/components/blog/source/shared/typing"
       ],
